+++ node/index.commonjs.js
@@ -1,5 +1,5 @@
 exports = module.exports = require('../commonjs/write/writeXlsxFileNode').default
 exports['default'] = require('../commonjs/write/writeXlsxFileNode').default
-exports.Integer = require('../commonjs/types/Integer').default
-exports.Email = require('../commonjs/types/Email').default
-exports.URL = require('../commonjs/types/URL').default
+// exports.Integer = require('../commonjs/types/Integer').default
+// exports.Email = require('../commonjs/types/Email').default
+// exports.URL = require('../commonjs/types/URL').default