+++ source/write/cell.js
@@ -1,6 +1,6 @@
-import Integer, { isInteger } from '../types/Integer'
-import URL, { isURL } from '../types/URL'
-import Email, { isEmail } from '../types/Email'
+// import Integer, { isInteger } from '../types/Integer'
+// import URL, { isURL } from '../types/URL'
+// import Email, { isEmail } from '../types/Email'
 
 import generateCellNumber from './generateCellNumber'
 import convertDateToExcelSerial from './convertDateToExcelSerial'
@@ -11,78 +11,29 @@ export default function generateCell(
   value,
   type,
   cellStyleId,
-  sharedStrings
+  getSharedString
 ) {
-  if (!isEmpty(value)) {
-    if (type === undefined) {
-      type = String
-      value = String(value)
-    }
-
-    // Available Excel cell types:
-    // https://github.com/SheetJS/sheetjs/blob/19620da30be2a7d7b9801938a0b9b1fd3c4c4b00/docbits/52_datatype.md
-    //
-    // Some other document (seems to be old):
-    // http://webapp.docx4java.org/OnlineDemo/ecma376/SpreadsheetML/ST_CellType.html
-    //
-    switch (type) {
-      case String:
-      case Email:
-      case URL:
-        if (type === Email && !isEmail(value)) {
-          throw new Error(`Invalid cell value: ${value}. Expected an Email`)
-        }
-        if (type === URL && !isURL(value)) {
-          throw new Error(`Invalid cell value: ${value}. Expected a URL`)
-        }
-        type = 's'
-        // // "inlineStr" type is used instead of "s" to avoid creating a "shared strings" index.
-        // type = 'inlineStr'
-        value = escapeString(value)
-        let id = sharedStrings.get(value)
-        if (id === undefined) {
-          id = sharedStrings.add(value)
-        }
-        value = id
-        break
-
-      case Number:
-      case Integer:
-        if (typeof value !== 'number') {
-          throw new Error(`Invalid cell value: ${value}. Expected a number`)
-        }
-        if (type === Integer && !isInteger(value)) {
-          throw new Error(`Invalid cell value: ${value}. Expected an Integer`)
-        }
-        // `n` is the default cell type (if no `t` has been specified).
-        type = undefined // 'n'
-        value = String(value)
-        break
-
-      case Date:
-        if (!(value instanceof Date)) {
-          throw new Error(`Invalid cell value: ${value}. Expected a Date`)
-        }
-        // "d" type doesn't seem to work.
-        // type = 'd'
-        // value = value.toISOString()
-        if (!cellStyleId) {
-          throw new Error('No "format" has been specified for a Date cell')
-        }
-        // `n` is the default cell type (if no `t` has been specified).
-        type = undefined // 'n'
-        value = convertDateToExcelSerial(value)
-        break
-
-      case Boolean:
-        type = 'b'
-        value = value ? '1' : '0'
-        break
-
-      default:
-        throw new Error(`Unknown schema type: ${type && type.name || type}`)
-    }
+  if (isEmpty(value)) {
+    return ''
   }
+
+  // Validate date format.
+  if (type === Date && !cellStyleId) {
+    throw new Error('No "format" has been specified for a Date cell')
+  }
+
+  // The default cell type is `String`.
+  if (type === undefined) {
+    type = String
+    // if (!isEmpty(value)) {
+      value = String(value)
+    // }
+  }
+
+  // if (!isEmpty(value)) {
+    value = getXlsxValue(type, value, getSharedString)
+  // }
+  type = getXlsxType(type)
 
   let cellStyle = ''
   // Available formatting style IDs (built-in in Excel):
@@ -90,15 +41,21 @@ export default function generateCell(
   // `2` — 0.00
   // `3` —  #,##0
   if (cellStyleId) {
+    // From the attribute s="12" we know that the cell's formatting is stored at the 13th (zero-based index) <xf> within the <cellXfs>
     cellStyle = ` s="${cellStyleId}"`
   }
 
   let valueType = ''
+  // The default value for `t` is `"n"` (a number or a date).
   if (type) {
     valueType = ` t="${type}"`
   }
 
-  return `<c r="${generateCellNumber(columnIndex, rowNumber)}"${valueType}${cellStyle}>${type === 'inlineStr' ? '<is><t>' : '<v>'}${isEmpty(value) ? '' : value}${type === 'inlineStr' ? '</t></is>' : '</v>'}</c>`
+  return `<c r="${generateCellNumber(columnIndex, rowNumber)}"${valueType}${cellStyle}>` +
+    (type === 'inlineStr' ? '<is><t>' : '<v>') +
+    (isEmpty(value) ? '' : value) +
+    (type === 'inlineStr' ? '</t></is>' : '</v>') +
+    '</c>'
 }
 
 /**
@@ -118,4 +75,84 @@ function escapeString(string) {
 
 function isEmpty(value) {
   return value === undefined || value === null || value === ''
+}
+
+function getXlsxType(type) {
+  // Available Excel cell types:
+  // https://github.com/SheetJS/sheetjs/blob/19620da30be2a7d7b9801938a0b9b1fd3c4c4b00/docbits/52_datatype.md
+  //
+  // Some other document (seems to be old):
+  // http://webapp.docx4java.org/OnlineDemo/ecma376/SpreadsheetML/ST_CellType.html
+  //
+  switch (type) {
+    case String:
+    // case Email:
+    // case URL:
+      return 's'
+      // // "inlineStr" type is used instead of "s" to avoid creating a "shared strings" index.
+      // return 'inlineStr'
+
+    case Number:
+    // case Integer:
+      // `n` is the default cell type (if no `t` has been specified).
+      // return 'n'
+      return
+
+    case Date:
+      // `n` is the default cell type (if no `t` has been specified).
+      // return 'n'
+      return
+
+    case Boolean:
+      return 'b'
+
+    default:
+      throw new Error(`Unknown schema type: ${type && type.name || type}`)
+  }
+}
+
+function getXlsxValue(type, value, getSharedString) {
+  // Available Excel cell types:
+  // https://github.com/SheetJS/sheetjs/blob/19620da30be2a7d7b9801938a0b9b1fd3c4c4b00/docbits/52_datatype.md
+  //
+  // Some other document (seems to be old):
+  // http://webapp.docx4java.org/OnlineDemo/ecma376/SpreadsheetML/ST_CellType.html
+  //
+  switch (type) {
+    case String:
+    // case Email:
+    // case URL:
+      // if (type === Email && !isEmail(value)) {
+      //   throw new Error(`Invalid cell value: ${value}. Expected an Email`)
+      // }
+      // if (type === URL && !isURL(value)) {
+      //   throw new Error(`Invalid cell value: ${value}. Expected a URL`)
+      // }
+      value = escapeString(value)
+      return getSharedString(value)
+
+    case Number:
+    // case Integer:
+      if (typeof value !== 'number') {
+        throw new Error(`Invalid cell value: ${value}. Expected a number`)
+      }
+      // if (type === Integer && !isInteger(value)) {
+      //   throw new Error(`Invalid cell value: ${value}. Expected an Integer`)
+      // }
+      return String(value)
+
+    case Date:
+      if (!(value instanceof Date)) {
+        throw new Error(`Invalid cell value: ${value}. Expected a Date`)
+      }
+      // "d" type doesn't seem to work.
+      // return value.toISOString()
+      return String(convertDateToExcelSerial(value))
+
+    case Boolean:
+      return value ? '1' : '0'
+
+    default:
+      throw new Error(`Unknown schema type: ${type && type.name || type}`)
+  }
 }