+++ source/write/cell.test.js
@@ -3,11 +3,9 @@
 
 import generateCell from './cell'
 
-import Integer from '../types/Integer'
-import Email from '../types/Email'
-import URL from '../types/URL'
-
-import SharedStrings from './sharedStrings'
+// import Integer from '../types/Integer'
+// import Email from '../types/Email'
+// import URL from '../types/URL'
 
 describe('generateCell()', () => {
   describe('Create a cell of type Sting', () => {
@@ -17,47 +15,47 @@ describe('generateCell()', () => {
     })
 
     it('should create a cell', () => {
-      const sharedStrings = new SharedStrings()
-      expect(generateCell(1, 0, 'Test', String, undefined, sharedStrings)).to.equal(
+      expect(generateCell(1, 0, 'Test', String, undefined, () => 0)).to.equal(
         // '<c r="A1" t="inlineStr"><is><t>Test</t></is></c>'
         '<c r="A1" t="s"><v>0</v></c>'
       )
-      expect(sharedStrings.sharedStrings).to.deep.equal(['Test'])
     })
   })
 
   describe('Create a cell of type Number', () => {
     it('should create a cell', () => {
-      expect(generateCell(1, 1, 1000, Number)).to.equal('<c r="B1"><v>1000</v></c>')
+      expect(generateCell(1, 1, 1000, Number))
+        .to.equal('<c r="B1"><v>1000</v></c>')
     })
+
     it('should create a cell with a cell style ID', () => {
-      expect(generateCell(1, 1, 1000, Number, 123)).to.equal('<c r="B1" s="123"><v>1000</v></c>')
+      const cellStyleId = 123
+      expect(generateCell(1, 1, 1000, Number, cellStyleId))
+        .to.equal(`<c r="B1" s="${cellStyleId}"><v>1000</v></c>`)
     })
   })
 
-  describe('Create a cell of type Integer', () => {
-    it('Create a cell', () => {
-      expect(generateCell(1, 1, 1000, Integer)).to.equal('<c r="B1"><v>1000</v></c>')
-    })
-  })
+  // describe('Create a cell of type Integer', () => {
+  //   it('Create a cell', () => {
+  //     expect(generateCell(1, 1, 1000, Integer)).to.equal('<c r="B1"><v>1000</v></c>')
+  //   })
+  // })
 
-  describe('Create a cell of type URL', () => {
-    it('should create a cell', () => {
-      const sharedStrings = new SharedStrings()
-      expect(generateCell(1, 1, 'https://google.com', URL, undefined, sharedStrings))
-        // .to.equal('<c r="B1" t="inlineStr"><is><t>https://google.com</t></is></c>')
-        .to.equal('<c r="B1" t="s"><v>0</v></c>')
-    })
-  })
+  // describe('Create a cell of type URL', () => {
+  //   it('should create a cell', () => {
+  //     expect(generateCell(1, 1, 'https://google.com', URL, undefined, () => 0))
+  //       // .to.equal('<c r="B1" t="inlineStr"><is><t>https://google.com</t></is></c>')
+  //       .to.equal('<c r="B1" t="s"><v>0</v></c>')
+  //   })
+  // })
 
-  describe('Create a cell of type Email', () => {
-    it('should create a cell', () => {
-      const sharedStrings = new SharedStrings()
-      expect(generateCell(1, 1, 'example@domain.com', Email, undefined, sharedStrings))
-        // .to.equal('<c r="B1" t="inlineStr"><is><t>example@domain.com</t></is></c>')
-        .to.equal('<c r="B1" t="s"><v>0</v></c>')
-    })
-  })
+  // describe('Create a cell of type Email', () => {
+  //   it('should create a cell', () => {
+  //     expect(generateCell(1, 1, 'example@domain.com', Email, undefined, () => 0))
+  //       // .to.equal('<c r="B1" t="inlineStr"><is><t>example@domain.com</t></is></c>')
+  //       .to.equal('<c r="B1" t="s"><v>0</v></c>')
+  //   })
+  // })
 
   describe('Create a cell of type Date', () => {
     it('should throw if no date format was supplied', () => {
@@ -66,10 +64,22 @@ describe('generateCell()', () => {
     })
 
     it('should create a cell', () => {
-      const sharedStrings = new SharedStrings()
-      expect(generateCell(1, 0, new Date(2020, 11, 30), Date, 123))
+      const cellStyleId = 123
+      expect(generateCell(1, 0, new Date(2020, 11, 30), Date, cellStyleId, () => 0))
         // .to.equal('<c r="B1" t="inlineStr"><is><t>example@domain.com</t></is></c>')
-        .to.equal('<c r="A1" s="123"><v>44194.875</v></c>')
+        .to.equal(`<c r="A1" s="${cellStyleId}"><v>44194.875</v></c>`)
+    })
+
+    it('should create an empty cell (`undefined`)', () => {
+      const cellStyleId = 123
+      expect(generateCell(1, 0, undefined, Date, cellStyleId, () => 0))
+        .to.equal('')
+    })
+
+    it('should create an empty cell (`null`)', () => {
+      const cellStyleId = 123
+      expect(generateCell(1, 0, null, Date, cellStyleId, () => 0))
+        .to.equal('')
     })
   })
 })