+++ source/write/row.test.js
@@ -15,8 +15,9 @@ describe('generateRow()', () => {
       ],
       0,
       {
-        formatStyles: {},
-        sharedStrings
+        getSharedString(string) {
+          return 0
+        }
       }
     )).to.equal(
       // '<row r="1"><c r="A1" t="inlineStr"><is><t>Test</t></is></c><c r="B1"><v>1000</v></c></row>'