+++ source/write/rows.test.js
@@ -11,7 +11,9 @@ describe('generateRows()', () => {
       value: 1000,
       type: Number
     }]], {
-    	sharedStrings: new SharedStrings()
+      getSharedString(string) {
+        return 0
+      }
     })).to.equal(
     	// '<row r="1"><c r="A1" t="inlineStr"><is><t>Test</t></is></c><c r="B1"><v>1000</v></c></row>'
     	'<row r="1"><c r="A1" t="s"><v>0</v></c><c r="B1"><v>1000</v></c></row>'