+++ source/write/writeXlsxFileBrowser.js
@@ -10,11 +10,11 @@ import rels from './statics/rels'
 import contentTypes from './statics/[Content_Types].xml'
 
 import generateWorksheet from './worksheet'
-import generateStyles from './styles'
-import SharedStrings from './sharedStrings'
+import initStyles from './styles'
+import initSharedStrings from './sharedStrings'
 
-export default function writeXlsxFile(data, { fileName, schema } = {}) {
-  return generateXlsxFile(data, { schema }).then((blob) => {
+export default function writeXlsxFile(data, { fileName, schema, columns } = {}) {
+  return generateXlsxFile(data, { schema, columns }).then((blob) => {
     if (fileName) {
       return FileSaver.saveAs(blob, fileName)
     }
@@ -28,20 +28,20 @@ export default function writeXlsxFile(data, { fileName, schema } = {}) {
  * "The reason if you want to send the excel file or store it natively on Cordova/capacitor app".
  * @return {Blob}
  */
-function generateXlsxFile(data, { schema }) {
+function generateXlsxFile(data, { schema, columns }) {
   const zip = new JSZip()
 
   zip.file('_rels/.rels', rels)
   zip.file('[Content_Types].xml', contentTypes)
 
-  const sharedStrings = new SharedStrings()
-  const { styles, formatStyles } = generateStyles(data, { schema })
-  const worksheet = generateWorksheet(data, { schema, formatStyles, sharedStrings })
+  const { getSharedStringsXml, getSharedString } = initSharedStrings()
+  const { getStylesXml, getStyle } = initStyles()
+  const worksheet = generateWorksheet(data, { schema, columns, getStyle, getSharedString })
 
   const xl = zip.folder('xl')
   xl.file('workbook.xml', workbookXML)
-  xl.file('styles.xml', styles)
-  xl.file('sharedStrings.xml', sharedStrings.getXml())
+  xl.file('styles.xml', getStylesXml())
+  xl.file('sharedStrings.xml', getSharedStringsXml())
   xl.file('_rels/workbook.xml.rels', workbookXMLRels)
   xl.file('worksheets/sheet1.xml', worksheet)