+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+1.0.3 / 18.06.2021
+==================
+
+* Fixed opening the generated `*.xlsx` files in MS Office 2007 Excel.
+
+* Added `align` attribute.
+
 1.0.0 / 17.06.2021
 ==================